move exported function into util file to conform to react-refresh rule
