+++ frontend/lib/src/components/core/Block/ElementNodeRenderer.tsx
@@ -81,7 +81,7 @@ import { Skeleton } from "~lib/components/elements/Skeleton"
 import TextElement from "~lib/components/elements/TextElement"
 import { ComponentInstance } from "~lib/components/widgets/CustomComponent"
 import { VegaLiteChartElement } from "~lib/components/elements/ArrowVegaLiteChart"
-import { getAlertElementKind } from "~lib/components/elements/AlertElement/AlertElement"
+import { getAlertElementKind } from "~lib/components/elements/AlertElement/utils"
 import Maybe from "~lib/components/core/Maybe"
 import { FormSubmitContent } from "~lib/components/widgets/Form"
 import Heading from "~lib/components/shared/StreamlitMarkdown/Heading"
